pyproject.toml
[tool.poetry]
name = "github-slack-pr-notifier"
version = "1.5"
description = ""
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
slack-sdk = "^3.36.0"
PyGithub = "^2.8.1"
python-dotenv = "^1.1.1"
regex = "^2025.9.1"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
poethepoet = "^0.37.0"
pytest = "^8.4.2"
pytest-pylint = "^0.21.0"
mypy = "^1.17.1"
pytest-mypy = "^1.0.1"

[tool.poe.tasks]
_black = "black ."
_isort = "isort ."
format-code = ["_isort", "_black"]

[tool.pytest.ini_options]
addopts = "--verbose --pylint --pylint-rcfile=pyproject.toml --mypy"
pythonpath = "src"

[tool.isort]
profile = "black"

[tool.black]
line-length = 150
extend-exclude = '''
/(
    tests
)/
'''

[tool.mypy]
strict = true
ignore_missing_imports = true
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true

[tool.pylint.messages_control]
disable = "all"
enable = """,
    unused-import,
    fixme,
    useless-object-inheritance,
    seless-import-alias,
    unused-variable,
    unused-argument,
    unexpected-keyword-arg,
    string,
    unreachable,
    invalid-name,
    logging-format-interpolation,
    logging-fstring-interpolation,
    unnecessary-pass,
    broad-except,
    bare-except,
    consider-using-from-import,
    raise-missing-from,
    no-else-return
    """
ignored_argument_names = "_.*|^ignored_|^unused_|args|kwargs"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
